fix: resolve bug in data type of Measurement statistics

+++ src/main/java/com/basereh/app/Average.java
@@ -5,11 +5,11 @@ import java.util.List;
 public class Average implements StatisticMeasure {
 
     @Override
-    public float calculate(List<Integer> nums) {
-        int sum = 0;
-        for (int num: nums) {
+    public float calculate(List<Float> nums) {
+        float sum = 0;
+        for (float num: nums) {
             sum += num;
         }
-        return (float)sum / nums.size();
+        return sum / nums.size();
     }
 }
